add getaccounts in store, service, handler so that client could know his accounts

+++ internal/store/account_store.go
@@ -16,33 +16,76 @@ func NewAccountStore(db *sql.DB) *AccountStore {
 
 func (s *AccountStore) CreateAccount(ctx context.Context, balance, userID int) (*model.Account, error) {
 	query := `INSERT INTO accounts (balance, user_id) VALUES ($1,$2) RETURNING id`
-	var id int
-	err := s.db.QueryRowContext(ctx, query, balance, userID).Scan(&id)
+	var accountID int
+	err := s.db.QueryRowContext(ctx, query, balance, userID).Scan(&accountID)
 	if err != nil {
 		return nil, err
 	}
-	return &model.Account{ID: id, UserID: userID, Balance: balance}, nil
+	return &model.Account{AccountID: accountID, UserID: userID, Balance: balance}, nil
 }
 
-func (s *AccountStore) GetAccount(ctx context.Context, id int) (*model.Account, error) {
-	query := `SELECT id, user_id, balance FROM accounts WHERE id=$1`
-	var acc model.Account
-	err := s.db.QueryRowContext(ctx, query, id).Scan(&acc.ID, &acc.UserID, &acc.Balance)
+func (s *AccountStore) GetAccounts(ctx context.Context, userID int) ([]model.Account, error) {
+	query := `SELECT id, user_id, balance FROM accounts WHERE user_id=$1`
+	var accounts []model.Account
+	rows, err := s.db.QueryContext(ctx, query, userID)
 	if err != nil {
 		return nil, err
 	}
-	return &acc, nil
+
+	/**
+		What if I read all data?
+		When you do:
+	    for rows.Next() {
+	      ...
+		}
+
+		until completion, many drivers will automatically consume the rest of the result set and make the connection reusable.
+
+		However, the Go documentation still recommends: defer rows.Close()
+		because:
+
+		You might return early.
+		A scan might fail.
+		Future code changes may skip reading all rows.
+		It's the standard safe pattern.
+	*/
+	defer rows.Close()
+
+	for rows.Next() {
+		var acc model.Account
+		err := rows.Scan(&acc.AccountID, &acc.UserID, &acc.Balance)
+		if err != nil {
+			return nil, err
+		}
+		accounts = append(accounts, acc)
+	}
+
+	if err := rows.Err(); err != nil {
+		return nil, err
+	}
+	return accounts, nil
 }
 
-func (s *AccountStore) UpdateAccount(ctx context.Context, id int, amount int) error {
+func (s *AccountStore) GetAccountByID(ctx context.Context, accountID int) (*model.Account, error) {
+	query := `SELECT id, user_id, balance FROM accounts WHERE id = $1`
+	var account model.Account
+
+	err := s.db.QueryRowContext(ctx, query, accountID).Scan(&account.AccountID, &account.UserID, &account.Balance)
+	if err != nil {
+		return nil, err
+	}
+	return &account, nil
+}
+
+func (s *AccountStore) UpdateAccount(ctx context.Context, accountID, userID int, amount int) error {
 	var query string
 	if amount < 0 {
-		query = `UPDATE accounts SET balance = balance + $1 WHERE id = $2 AND balance + $1 >=0`
+		query = `UPDATE accounts SET balance = balance + $1 WHERE id = $2 AND user_id = $3 AND balance + $1 >=0`
 	} else {
-		query = `UPDATE accounts SET balance = balance + $1 WHERE id = $2`
+		query = `UPDATE accounts SET balance = balance + $1 WHERE id = $2 AND user_id = $3`
 	}
 
-	result, err := s.db.ExecContext(ctx, query, amount, id)
+	result, err := s.db.ExecContext(ctx, query, amount, accountID, userID)
 	if err != nil {
 		return err
 	}
@@ -108,10 +151,10 @@ func (s *AccountStore) TransferTx(ctx context.Context, fromID, toID, amount int)
 	return tx.Commit()
 }
 
-func (s *AccountStore) DeleteAccount(ctx context.Context, id int) error {
-	query := `DELETE FROM accounts WHERE id=$1`
+func (s *AccountStore) DeleteAccount(ctx context.Context, accountID, userID int) error {
+	query := `DELETE FROM accounts WHERE id=$1 AND user_id = $2`
 
-	result, err := s.db.ExecContext(ctx, query, id)
+	result, err := s.db.ExecContext(ctx, query, accountID, userID)
 	if err != nil {
 		return err
 	}
